move mentionset to message_editor
